refactor: renamed unsafe2->hack
